+++ tardis/resources/drone.py
@@ -69,7 +69,11 @@ class Drone(Pool):
         self._demand = value
 
     @property
-    def drone_minimum_lifetime(self) -> [int, None]:
+    def heartbeat_interval(self) -> int:
+        return self.site_agent.drone_heartbeat_interval
+
+    @property
+    def minimum_lifetime(self) -> [int, None]:
         return self.site_agent.drone_minimum_lifetime
 
     @property
@@ -98,7 +102,7 @@ class Drone(Pool):
                 )
                 self._demand = 0
                 return
-            await asyncio.sleep(60)
+            await asyncio.sleep(self.heartbeat_interval)
 
     def register_plugins(self, observer: Union[List[Plugin], Plugin]) -> None:
         self._plugins.append(observer)